✨ Allow configuration of serialized fields
Part of the reason for originally forking this library was to allow us
to specify extra fields in the `Delta` serialized JSON representation
when the document is written to the database.

We currently use this just for the `metadata` field, where we store
auth information about the owner and book IDs.

We're soon going to add `rich-text` migration to `reedsy-sharedb`, which
requires us to record if a document has been migrated, which we'll store
in a similar way.

Rather than have to change this library every time we want to serialize
a new field, this change instead adds module-level config, where
consumers can declare the fields they want (de)serialized:

```js
var json0 = require('@reedsy/rich-text')
json0.config.serializedProperties.extra = true;
```

Any fields that aren't declared here will be ignored when
(de)serializing.

By default we enable `metadata` in order to keep this change backwards-
compatible.